stroke_rehab/requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
opencv-python>=4.8
mediapipe>=0.10
pygame>=2.5
streamlit>=1.30
plotly>=5.18
pyserial>=3.5
bleak>=0.21         # BLE Bluetooth
joblib>=1.3
Pillow>=10.0
